requirements.txt
markdown2==1.0.1.17
Django==1.2.3
